requirements.txt
pinecone-client
langchain
langchain-community
langchain-pinecone
langchain-groq
langchain-core
streamlit
python-dotenv
sentence-transformers
pypdf
faiss-cpu